requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bitcoin==1.1.42
bitmerchant==0.1.8
black==22.6.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coincurve==17.0.0
commontools==0.1.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.17.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
ipykernel==6.9.2
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keychain==0.14.2.0
keylib==0.1.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
playwright==1.25.2
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==4.21.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybitcoin==0.9.9
pycparser==2.21
pyee==8.1.0
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-bitcoinrpc==0.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.9
utilitybelt==0.2.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
widgetsnbextension==3.6.0
